Four constants have been declared for use in the bookmarklet. These constants are:
siteUrl and staticUrl: the base URL of the website and the base URL of static files.
minWidth and minHeight: the minimum width and height of the images that the bookmarklet will fetch from the website, in pixels. The bookmarklet will identify images that have a width and height of at least 250px each.